+++ CHANGES.md
@@ -17,9 +17,10 @@ Beta Releases
    * Replaced `Uniform.getFrameNumber` and `Uniform.getTime` with `Uniform.getFrameState`, which returns the full frame state.    
 * Improved the performance of drawing polygons created with `configureFromPolygonHierarchy`.
 * Added `GeometryFilters.combine` to combine meshes for better batching.
+* Added `GeometryFilters.computeNormals` to find the normals of vertices in a mesh.
 * Added `BoundingSphere.fromEllipsoid`.
 * Added renderer support for `OES_element_index_uint`, which can improve performance by reducing batch sizes.
-* Added 'czm_tangentToEyeSpaceMatrix` built-in GLSL function.
+* Added `czm_tangentToEyeSpaceMatrix` built-in GLSL function.
 
 ### b17 - 2013-06-03